fix: encode login credentials in UTF-8 form for non-ascii character compatibility
Currently we use default UCS-2 to encode login credentials in base64 form.
But in signup form, the codec we are using is UTF-8, which result to a different
byte sequence for non-ascii characters and cause login failure.

By changing the login form codec to UTF-8, this issue can be fixed and has no
impact to any existing user.